perf: Optimize SQL queries to eliminate N+1 problems
Eliminated multiple N+1 query issues in backend:

1. Hashtag insertion optimization:
   - Before: 3 queries per hashtag in a loop (3N queries)
   - After: Batch insert using UNNEST (3 queries total)
   - Affected: create_operation, update_operation handlers
   - New helper: link_operation_hashtags() for batch processing

2. List operations optimization:
   - Before: Query hashtags per operation in a loop (N queries)
   - After: Single batch query for all operations (1 query)
   - Affected: list_operations handler
   - New helper: get_operations_hashtags_batch() using ANY($1)

3. Type fixes:
   - Changed CreateOperation.amount from f64 to BigDecimal
   - Ensures consistency with database NUMERIC type

Performance impact:
- Creating operation with 3 hashtags: 9 queries → 3 queries (67% reduction)
- Listing 20 operations with hashtags: 21 queries → 2 queries (90% reduction)

Tested and verified working with curl commands.

+++ backend/src/models.rs
@@ -166,7 +166,7 @@ pub struct CreateOperation {
     pub category_id: Option<i32>,
     pub description: Option<String>,
     pub asset_id: i32,
-    pub amount: f64,
+    pub amount: BigDecimal,
     pub operation_type: String,
     pub operation_date: String,
 }
